gradio_ui/requirements.txt
gradio>=4.0.0,<5.0.0 # Use a recent version of Gradio
requests>=2.28.0
python-dotenv>=1.0.0
